add permission control for private posts && admin

+++ app/Http/Controllers/Admin/PostController.php
@@ -15,12 +15,18 @@ class PostController extends Controller
      */
     public function index()
     {
-        // $posts = Post::latest()->paginate(10);
-        $posts = Post::with('user')->with('category')->where('private', '=', 'false')->paginate(10);
-        $user = auth()->user();
-        // $posts = User::all();
-        // dd($posts);
-        return view('admin.posts.index', compact('posts', 'user'));
+        $current_user = auth()->user();
+        if(count($current_user->roles) > 0){
+            if($current_user->roles[0]->id <= 2){
+                // $posts = Post::latest()->paginate(10);
+                $posts = Post::with('user')->with('category')->where('private', '=', 'false')->paginate(10);
+                $user = auth()->user();
+                // $posts = User::all();
+                // dd($posts);
+                return view('admin.posts.index', compact('posts', 'user'));
+            }
+        } 
+        return view('errors.noauth');
     }
 
     /**
@@ -30,7 +36,13 @@ class PostController extends Controller
      */
     public function create()
     {
-        //
+        $current_user = auth()->user();
+        if(count($current_user->roles) > 0){
+            if($current_user->roles[0]->id <= 2){
+                //
+            }
+        } 
+        return view('errors.noauth');
     }
 
     /**
@@ -41,7 +53,13 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $current_user = auth()->user();
+        if(count($current_user->roles) > 0){
+            if($current_user->roles[0]->id <= 2){
+                //
+            }
+        } 
+        return view('errors.noauth');
     }
 
     /**
@@ -52,7 +70,13 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        //
+        $current_user = auth()->user();
+        if(count($current_user->roles) > 0){
+            if($current_user->roles[0]->id <= 2){
+                //
+            }
+        } 
+        return view('errors.noauth');
     }
 
     /**
@@ -63,19 +87,37 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        //
+        $current_user = auth()->user();
+        if(count($current_user->roles) > 0){
+            if($current_user->roles[0]->id <= 2){
+                //
+            }
+        } 
+        return view('errors.noauth');
     }
 
     public function publish(Post $post){
-        $post->published = true;
-        $post->save();
-        return redirect(route('admin.posts.index'));
+        $current_user = auth()->user();
+        if(count($current_user->roles) > 0){
+            if($current_user->roles[0]->id <= 2){
+                $post->published = true;
+                $post->save();
+                return redirect(route('admin.posts.index'));
+            }
+        } 
+        return view('errors.noauth');
     }
 
     public function unpublish(Post $post){
-        $post->published = false;
-        $post->save();
-        return redirect(route('admin.posts.index'));
+        $current_user = auth()->user();
+        if(count($current_user->roles) > 0){
+            if($current_user->roles[0]->id <= 2){
+                $post->published = false;
+                $post->save();
+                return redirect(route('admin.posts.index'));
+            }
+        } 
+        return view('errors.noauth');
     }
 
     /**
@@ -87,7 +129,13 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        //
+        $current_user = auth()->user();
+        if(count($current_user->roles) > 0){
+            if($current_user->roles[0]->id <= 2){
+                //
+            }
+        } 
+        return view('errors.noauth');
     }
 
     /**
@@ -98,7 +146,13 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        $post->delete();
-        return redirect(route('admin.posts.index'));
+        $current_user = auth()->user();
+        if(count($current_user->roles) > 0){
+            if($current_user->roles[0]->id <= 2){
+                $post->delete();
+                return redirect(route('admin.posts.index'));
+            }
+        } 
+        return view('errors.noauth');
     }
 }
